Add permissionsOf(memberID) to [GuildChannel] and [GuildTextableChannel] and added `stageModerator` permissions
